// frontend/src/company.d.ts
export interface Schedule {
  ScheduleID: number;
  DoctorID: number;
  Time: string;
  Day: string;
}